test: refactor fetch users test
