+++ src/http/controllers/health/health-controller.ts
@@ -1,5 +1,5 @@
 import { SuccessResponse } from "@/@types/response";
-import { env } from "../../env/index";
+import { env } from "@/env";
 import { Request, Response } from "express";
 
 // Controller para health check detalhado
